Tidied ThreadArgs struct and began separating code
Used difference definition of struct so that struct doesn't have to be
written everywhere. Also changed the way struct is instantiated, but it
is possible this will not compile on Balena due to it's version of C and
it's compiler, so left the old version untouched.

Began separating code into separate c/h files. Currently have only done
this for debuging functions. To be continued.