WIP. Changed count to array of available threads
This now works by building an array of flags that indicate which thread
is and isn't available. A corresponding array of locks is also built.
For each flag the lock is acquired and the value updated to signify
whether or not the thread is available or not.

This allows the program to wait for all the threads to finish correctly
as this was the source of a bug before, where there was no way of
knowing how many threads to wait for.

Minimal testing done so far and lots of cleaning up needed but initially
looking like this works. To be continued.

NOTE: This is the first commit of this branch, branched off of
code-formatting-and-organisation branch. Merge back into this branch
then to master.