Added CLI tool with help and fixed problems
Five set problems were randomly generated and have been stored in
problems.c with functions to fill an input array with the values from
them.
This now allows the user to specify:
 - Problem id
 - Threads to use
 - Precision to work to
The user can now also use --help or -h options.
This will be useful during testing so that problems are consistent and
therefore testing is fair.