Fixed bug by correcting check before pthread_join
This was skipping threads if they were not available, where it should
have been skipping them if they were available. If a thread is marked as
available then we know no thread is running with that ID. If a thread is
marked as unavailable then we know a thread is running with that ID, so
we should pthread_join it.