Fixed bug and added comments
The bug was due to three issues:
1. The wereValuesModified flag system didn't work, as this was checked
potentially before it was updated. Instead, the resetting of the
valuesSolvedArray has been moved into the threads themselves. The thread
still only updates the one corresponding solved flag if it sees that a
value is solved, but now resets the entire valuesSolvedArray if it
updates a value.
2. The while loop was terminating when the valuesSolvedArray contained
no 0s. However, a thread could then see a value changed and reset the
array. Instead this loop now only terminates when all threads have
finished, AND the valuesSolvedArray contains no 0s.
3. The struct passed to each thread was actually the same struct each
time. This instead needed to be done using malloc. This revealed a
bigger issue where this algorithm was very quickly being reduced from
the given number of threads to just one, as all threads were just
setting the last flag in the threadsAvailable array to 1, rather than
their own flag. The struct is now malloced, passed in, and freed in
updateValueProxy before it returns.

Also added lots of comments and made some readability changes. This file
is still quite unreadable and is full of comments to try to help with
this. Would like to improve it's readability further but not sure how to
do this without sacrificing some speed of the code.