Fixed this very difficult bug
When running on balena with problem 5 and precision <0.3, the program
stopped after a while and never terminated. This turned out to be due to
the fact that no new threads could be created. This was because threads
were not 'pthread_detach'ed, meaning memory was not released into the
system when threads terminated. This meant after a while the memory
limit was exceeded and the OS rejected the creation of any more threads.

The actual issue was because thread creation failed, but the thread
availability flag was still set to 0. This leads to the note below.

NOTE: NEED TO CHECK ERROR RESPONSE CODES AS THIS WOULD HAVE MASSIVELY
HELPED WITH THIS. ADD CHECKS TO ALL PTHREAD_* FUNCTIONS!